Add exception handler to stop
This is to make our tests run by handling an edge case exception.

One of our tests starts and stop a BlockTracker without interracting
with web3. When we stop, this results in an exception because the
eth-block-tracker-es5 v2.32 checks for a web3 subscription before
unwinding from the saga-channel unsubscripte.

Ref: node_modules/eth-block-tracker-es5/lib/index.js:177
Note - this module is ancient and github repo shows it is currently at
v4.0, under a new name. Drizzle seems to be locked into a release
persisted by npm registry.